requirements.txt
pyrogram
numpy
Pillow
wikipedia
wordcloud
tgcrypto
